ci: add MassSpring3D test

+++ test/system_test/MassSpring3D/Shader.cpp
@@ -24,7 +24,7 @@ GLShader::~GLShader() {
 void GLShader::compile(const char* source) {
 	GLint compiled = 0;  // Compiled flag
 	const char* ptrs[] = { source };
-	const GLint lens[] = { std::strlen(source) };
+	const GLint lens[] = { static_cast<GLint>(std::strlen(source)) };
 	glShaderSource(handle, 1, ptrs, lens);
 	glCompileShader(handle);
 	glGetShaderiv(handle, GL_COMPILE_STATUS, &compiled);
